backend/korevitta/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
dj-database-url>=2.2,<3.0
psycopg[binary]>=3.2,<4.0
